partial code moved
