+++ src/gen/mod.rs
@@ -1,4 +1,107 @@
-// ... (previous content remains the same)
+use std::collections::HashMap;
+
+use anyhow::Result;
+use camino::Utf8Path;
+
+use genco::fmt;
+use genco::prelude::*;
+use serde::{Deserialize, Serialize};
+// use uniffi_bindgen::MergeWith;
+use uniffi_bindgen::{BindingGenerator, BindingsConfig, ComponentInterface};
+
+use self::render::Renderer;
+use self::types::TypeHelpersRenderer;
+
+mod compounds;
+mod enums;
+mod functions;
+mod objects;
+mod oracle;
+mod primitives;
+mod records;
+mod render;
+mod types;
+mod utils;
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+pub struct Config {
+    package_name: Option<String>,
+    cdylib_name: Option<String>,
+    #[serde(default)]
+    external_packages: HashMap<String, String>,
+}
+
+// impl MergeWith for Config {
+//     fn merge_with(&self, other: &Self) -> Self {
+//         let package_name = if other.package_name.is_some() {
+//             other.package_name.clone()
+//         } else {
+//             self.package_name.clone()
+//         };
+//         let cdylib_name = if other.cdylib_name.is_some() {
+//             other.cdylib_name.clone()
+//         } else {
+//             self.cdylib_name.clone()
+//         };
+//         Self {
+//             package_name,
+//             cdylib_name,
+//         }
+//     }
+// }
+
+impl From<&ComponentInterface> for Config {
+    fn from(ci: &ComponentInterface) -> Self {
+        Config {
+            package_name: Some(ci.namespace().to_owned()),
+            cdylib_name: Some(ci.namespace().to_owned()),
+            external_packages: HashMap::new(),
+        }
+    }
+}
+
+impl Config {
+    pub fn package_name(&self) -> String {
+        if let Some(package_name) = &self.package_name {
+            package_name.clone()
+        } else {
+            "uniffi".into()
+        }
+    }
+
+    pub fn cdylib_name(&self) -> String {
+        if let Some(cdylib_name) = &self.cdylib_name {
+            cdylib_name.clone()
+        } else {
+            "uniffi".into()
+        }
+    }
+}
+
+impl BindingsConfig for Config {
+    fn update_from_ci(&mut self, ci: &ComponentInterface) {
+        self.package_name = Some(ci.namespace().to_owned());
+    }
+
+    fn update_from_cdylib_name(&mut self, cdylib_name: &str) {
+        self.cdylib_name = Some(cdylib_name.to_string());
+    }
+
+    fn update_from_dependency_configs(&mut self, config_map: HashMap<&str, &Self>) {
+        for (crate_name, config) in config_map {
+            if !self.external_packages.contains_key(crate_name) {
+                self.external_packages
+                    .insert(crate_name.to_string(), config.package_name());
+            }
+        }
+    }
+}
+
+pub struct DartWrapper<'a> {
+    config: &'a Config,
+    // ci: &'a ComponentInterface,
+    type_renderer: TypeHelpersRenderer<'a>,
+}
 
 impl<'a> DartWrapper<'a> {
     // ... (previous methods remain the same)
